Feedback-Add cancel button to forms-pt159251141
There was a need for users to be able to cancel and go back if they did not want to update or create an entry, or update their profile. So a cancel button has been added to each of the forms to allow a user to leave the form and return to the previous page.