Feature- Create directory structure-#159428290
This commit creates the directory structure that will be used to build the remaining part of the application connecting the front-end to the api.
A new folder ‘FULLCONNECTION’ has been made in which all the files for this new feature and later features will be built.
Babel and eslint with the recommended style guide has also been setup.
Node packages have also been installed.

Feature: #159428290