Feature-Sign up to api endpoint-#159428290
This commit adds the connection between the api endpoint for the signup and the front-end page. This allows a user to be able to signup by submitting his/her details to the form.
Upon successful registration, the user is redirected to the signup page to signin.
Some other changes were also made such as the creation of a proper dialog box for user feedback.

The following files were changed:
- javascripts/es6/scripts.js
- stylesheets/forms.css
- stylesheets/global.css
- GruntFile.js
- package.json
- .gitignore
- signup.html
- signin.html

Feature: #159428290