ft-Modify signup page UI-#159428290
This commit adds a maximum date which a user can set as a date of birth. This commit also removes the dashboard and about link from the footer because they were not needed. The buttons for signup and cancel have been modified to be of the same width.
The following files were changed:
- signup.html
- signin.html
- signup.css
- global.css

ft-#159428290