ft-create sign in api connection to ui-#159428325
This commit connects the sign in api endpoint to the signin page. This allows a user to be able to log into the app.
The footer has also been changed for all other files to be uniform.
Scss files have been removed from .gitignore.
The following files were added/changed:
- signin.html
- stylesheets/signup.css
- stylesheets/*.scss
- all html pages

Ft: #159428325