ft-add logout and page authentication-#159478300
This commit adds the functionality to allow a user to logout of the application when he/she is done. This commit also ensures that an unauthenticated user can only access the signup and signin routes. The navigation links now show depending on if a user is signed in or not.
There has also been small changes in the styling of certain pages.
The dashboard now gets the last 4 entries of a user and displays it.

Ft: #159478300, 159478352